One ParseExtension object per parser
