+++ src/main/scala/viper/silver/ast/utility/Consistency.scala
@@ -29,7 +29,7 @@ object Consistency {
     recordIfNot(suspect, !property, message)
 
   /** Names that are not allowed for use in programs. */
-  def reservedNames: Seq[String] = FastParserCompanion.keywords.toSeq
+  def reservedNames: Seq[String] = FastParserCompanion.basicKeywords.toSeq
 
   /** Returns true iff the string `name` is a valid identifier. */
   val identFirstLetter = "[a-zA-Z$_]"
